pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "arcana-changeme"
description = "Description for arcana-changeme goes here"
readme = "README.rst"
requires-python = ">=3.8"
dependencies = [
    "arcana >=0.9.4"
]
license = {file = "LICENSE"}
authors = [{name = "Change Me", email = "changeme@changeme.com"}]
maintainers = [{name = "Change Me", email = "changeme@changeme.com"}]
keywords = ["arcana"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "black",
    "codespell",
    "flake8",
    "flake8-pyproject",
    "pre-commit",
]
doc = [
    "docutils>=0.10",
    "furo>=2022.2.14.1",
    "mock>1.0",
    "numpydoc>=0.6.0",
    "packaging",
    "sphinx >=2.1.2",
    "sphinx-argparse>=0.2.0",
    "sphinx-click>=3.1",
]
test = [
    "pytest >= 4.4.0",
    "pytest-cov",
    "pytest-env",
    "codecov",
]

[project.urls]
documentation = "https://arcana.readthedocs.io"
homepage = "https://arcana.readthedocs.io"
# repository = "https://github.com/ArcanaFramework/arcana-changeme.git"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "arcana/changeme/_version.py"

[tool.hatch.build]
packages = [
    "arcana",
]
exclude = [
    "/tests",
]

[tool.black]
line-length = 88
target-version = ["py38"]
exclude = "_version.py"

[tool.codespell]
ignore-words = ".codespell-ignorewords"
